missing routes frienship and crew

+++ script/sql/fonction/user.sql
@@ -76,19 +76,30 @@ $$ LANGUAGE plpgsql SECURITY DEFINER;
 
 -------------------------------------------------------------------------------------------------- !
 -- fonction qui supprime un utilisateur et retourne true ou false
-CREATE OR REPLACE FUNCTION web.delete_user(user_id int)
+CREATE OR REPLACE FUNCTION web.delete_user(p_user_id int)
 RETURNS BOOLEAN
 AS $$
 BEGIN
+    -- Delete related records in web.r_user_event
+    DELETE FROM web.r_user_event WHERE web.r_user_event.user_id = p_user_id;
+
+    -- Delete related records in web.r_user_crew
+    DELETE FROM web.r_user_crew WHERE web.r_user_crew.user_id = p_user_id;
+
+    -- Delete crews created by the user
+    DELETE FROM web.crew WHERE web.crew.user_id = p_user_id;
+
+    -- Delete links with other users in web.contact
+    DELETE FROM web.contact WHERE web.contact.user_id = p_user_id OR web.contact.friend_id = p_user_id;
+
     -- Check if the user exists
     IF EXISTS (
         SELECT 1
         FROM main.user
-        WHERE id = user_id
+        WHERE id = p_user_id
     ) THEN
         -- Delete the user
-        DELETE FROM main.user
-        WHERE id = user_id;
+        DELETE FROM main.user WHERE id = p_user_id;
 
         -- Return true indicating successful deletion
         RETURN true;
@@ -98,3 +109,5 @@ BEGIN
     END IF;
 END;
 $$ LANGUAGE plpgsql SECURITY DEFINER;
+
+
